Add error handling for delete endpoint when id not in database

+++ src/mortgages/mortgages.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Delete, Body, ParseIntPipe, Param } from '@nestjs/common';
+import { Controller, Post, Delete, Body, ParseIntPipe, Param, NotFoundException } from '@nestjs/common';
 import { MortgagesService } from './mortgages.service';
 import { CreateMortgageDto } from './dtos/CreateMortgage.dto';
 
@@ -12,7 +12,11 @@ export class MortgagesController {
   }
 
   @Delete(':id')
-  async deleteMortgageById(@Param('id', ParseIntPipe) id: number) {
+  async deleteMortgageById(@Param('id', ParseIntPipe) id: number): Promise<void> {
+    const mortgage = await this.mortgagesService.findOneById(id)
+    if (!mortgage) {
+      throw new NotFoundException(`Entity with ID ${id} not found`)
+    }
     await this.mortgagesService.deleteMortgage(id);
   }
 }
